fix: Updating UAT tests workflow and moving terra tests under one file
