+++ .github/workflows/terratests-uat-suite.yaml
@@ -10,14 +10,6 @@ jobs:
     runs-on: ubuntu-latest
     env:
       EQUINIX_API_ENDPOINT: "https://uatapi.equinix.com"
-      TEST_DATA_UAT_PORT_2_PORT_CONNECTION: ${{secrets.TEST_DATA_UAT_PORT_2_PORT_CONNECTION}}
-      TEST_DATA_UAT_PORT_2_AZURE_CONNECTION: ${{secrets.TEST_DATA_UAT_PORT_2_AZURE_CONNECTION}}
-      TEST_DATA_UAT_PORT_2_AWS_CONNECTION: ${{secrets.TEST_DATA_UAT_PORT_2_AWS_CONNECTION}}
-      TEST_DATA_UAT_PORT_2_IBM2_CONNECTION: ${{secrets.TEST_DATA_UAT_PORT_2_IBM2_CONNECTION}}
-      TEST_DATA_UAT_PORT_2_ALIBABA_CONNECTION: ${{secrets.TEST_DATA_UAT_PORT_2_ALIBABA_CONNECTION}}
-      TEST_DATA_UAT_CLOUD_ROUTER_2_PORT_CONNECTION: ${{secrets.TEST_DATA_UAT_CLOUD_ROUTER_2_PORT_CONNECTION}}
-      TEST_DATA_UAT_PORT_2_PUBLIC_SERVICE_PROFILE_CONNECTION: ${{secrets.TEST_DATA_UAT_PORT_2_PUBLIC_SERVICE_PROFILE_CONNECTION}}
-      TEST_DATA_UAT_PORT_2_PRIVATE_SERVICE_PROFILE_CONNECTION: ${{secrets.TEST_DATA_UAT_PORT_2_PRIVATE_SERVICE_PROFILE_CONNECTION}}
 
     steps:
       - name: Check out code into the Go module directory
@@ -41,30 +33,22 @@ jobs:
 
       - name: Setup Variables Files
         run: |
-          for list in ./examples/*; do
-              if [[ "$list" == *"port-2-port-connection"* ]]; then
-                  echo $TEST_DATA_UAT_PORT_2_PORT_CONNECTION >> "$list/terraform.tfvars.json"
-              elif [[ "$list" == *"port-2-ibm2-connection"* ]]; then
-                  echo $TEST_DATA_UAT_PORT_2_IBM2_CONNECTION >> "$list/terraform.tfvars.json"
-              elif [[ "$list" == *"port-2-alibaba-connection"* ]]; then
-                  echo $TEST_DATA_UAT_PORT_2_ALIBABA_CONNECTION >> "$list/terraform.tfvars.json"
-              elif [[ "$list" == *"port-2-public-service-profile-connection"* ]]; then
-                  echo $TEST_DATA_UAT_PORT_2_PUBLIC_SERVICE_PROFILE_CONNECTION >> "$list/terraform.tfvars.json"
-              elif [[ "$list" == *"port-2-private-service-profile-connection"* ]]; then
-                  echo $TEST_DATA_UAT_PORT_2_PRIVATE_SERVICE_PROFILE_CONNECTION >> "$list/terraform.tfvars.json"
-              fi     
-          done
-          for list in ./tests/uat/*; do
-              if [[ "$list" == *"port-2-azure-connection"* ]]; then
-                  echo $TEST_DATA_UAT_PORT_2_AZURE_CONNECTION >> "$list/terraform.tfvars.json"
-              elif [[ "$list" == *"port-2-aws-connection"* ]]; then
-                  echo $TEST_DATA_UAT_PORT_2_AWS_CONNECTION >> "$list/terraform.tfvars.json"
-              fi
-              
-          done
+          echo ${{secrets.TEST_DATA_UAT_PORT_2_ALIBABA_CONNECTION}} >> "./examples/port-2-alibaba-connection/terraform.tfvars.json"
+          echo ${{secrets.TEST_DATA_UAT_PORT_2_AWS_CONNECTION}} >> "./tests/uat/port-2-aws-connection/terraform.tfvars.json"
+          echo ${{secrets.TEST_DATA_UAT_PORT_2_AZURE_CONNECTION}} >> "./tests/uat/port-2-azure-connection/terraform.tfvars.json"
+          echo ${{secrets.TEST_DATA_UAT_PORT_2_IBM2_CONNECTION}} >> "./examples/port-2-ibm2-connection/terraform.tfvars.json"
+          echo ${{secrets.TEST_DATA_UAT_PORT_2_PORT_CONNECTION}} >> "./examples/port-2-port-connection/terraform.tfvars.json"
+          echo ${{secrets.TEST_DATA_UAT_PORT_2_PRIVATE_SERVICE_PROFILE_CONNECTION}} >> "./examples/port-2-private-service-profile-connection/terraform.tfvars.json"
+          echo ${{secrets.TEST_DATA_UAT_PORT_2_PUBLIC_SERVICE_PROFILE_CONNECTION}} >> ".examples/port-2-public-service-profile-connection/terraform.tfvars.json"   
+
       - name: Run Go Tests
         run:
           go test ./tests/uat -v -coverprofile coverage_uat_modules.txt -covermode=atomic -count 1 -parallel 8 -timeout 180m
+
+      - name: Setup Node.js 20.11.1
+        uses: actions/setup-node@v2
+        with:
+            node-version: 20.11.1
 
       - name: Upload test coverage to Codecov
         if: ${{ always() }}
