+++ tests/uat/port-2-azure-connection/main.tf
@@ -1,7 +1,6 @@
 provider "equinix" {
   client_id     = var.equinix_client_id
   client_secret = var.equinix_client_secret
-  endpoint = "https://uatapi.equinix.com"
 }
 module "create_port_2_azure_connection" {
   source = "equinix/fabric/equinix//modules/port-connection"
