--- tests/uat/port-2-port-connection_test.go
@@ -1,21 +1,0 @@
-package uat
-
-import (
-	"github.com/gruntwork-io/terratest/modules/terraform"
-	"github.com/stretchr/testify/assert"
-	"testing"
-)
-
-func TestPort2PortCreateConnection(t *testing.T) {
-	// retryable errors in terraform testing.
-
-	terraformOptions := terraform.WithDefaultRetryableErrors(t, &terraform.Options{
-		TerraformDir: "../../examples/port-2-port-connection",
-	})
-
-	defer terraform.Destroy(t, terraformOptions)
-
-	terraform.InitAndApply(t, terraformOptions)
-	output := terraform.Output(t, terraformOptions, "port_connection_id")
-	assert.NotNil(t, output)
-}
